update extension functions and reformat
