+++ core/data/src/test/java/com/veritalex/core/data/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package com.veritalex.core.data
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
